// src/types/generated/ProcessInfo.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type ProcessInfo = { pid: number, ppid: number, user: string, state: string, cpuPercent: number, memPercent: number, rssKb: number, elapsedSecs: number, name: string, command: string, };